vo gl: Output complete suboption help
The output is incomplete because of the buffer in mp_msg
implementation being smaller than the help output.

Fixing it by increasing the maximal message length in
mp_msg.c seems not wise, as usually 2 buffers of that
length are allocated on the stack.

So just print help with 2 mp_msg calls.


git-svn-id: svn://svn.mplayerhq.hu/mplayer/trunk@37562 b3059339-0415-0410-9bf9-f77b7e298cf2