Prepare cue sheet playlist support in playlist dialog.
Add complete original plItem information to the file name and path
which have been converted for display.

Retrieve and use this original information when rebuilding the playlist
(i.e. when appending CLSelected rows).

Playlist items to be newly added can only be real files (the contents
of cue sheet files will be parsed later), so it's enough to add only
original file name and path in this case.


git-svn-id: svn://svn.mplayerhq.hu/mplayer/trunk@37982 b3059339-0415-0410-9bf9-f77b7e298cf2