Make Win32 GUI's widget renderer more compatible
The X11/GTK GUI renderer will not draw anything, if the image generated
from a label is empty.

The Win32 GUI renderer always repaints the background of a label image
prior to drawing the image generated from the label itself. Prevent the
renderer from doing this by checking whether there actually will be any
text to be rendered at all. (This is no 100% compatibility, but good
enough for existing skins.)


git-svn-id: svn://svn.mplayerhq.hu/mplayer/trunk@37989 b3059339-0415-0410-9bf9-f77b7e298cf2